Re-stupiding some smart directory shuffling
To me it makes more sense to have views in the client directory, and server/routes in the server directory.
However, this made all my pages disappear, and I can't be arsed to figure out what changes need to be made in code
to make those directory moves not break shit.