Refactor for edn-readers with an inventory handler.
I hope I'm getting the hang of how to use readers. When you pass data off (*write*) you're packing it into the best
structure you can (generally we're going for "simpler is gooder" here), writing an address on the envelope of data,
putting it into a box with who knows what other data. When you pick data up (*read*), you might want to do all kinds of
things with it, and so data is handed to a particular namespace and name.

See data_readers.clj for an idea of where I'm going with this.