Mongo is persisting and I can connect to it!
I really though that I was going to be doing weird things with serializeable EDN format. Instead,
I'm all set up with a local db which NM tells me will not be so difficult to change over using fancy tunneling.
So I'm going to treat this as a db and do it to it.